Correct some bug

+++ upload.php
@@ -20,38 +20,46 @@ if (isset($_FILES['file'])){
 
 	$name = $_FILES['file']['name'];
 	$temp_name = $_FILES['file']['tmp_name'];
+	$type = $_FILES['file']['type'];
     $array = explode('.', $_FILES['file']['name']);
     $extension = strtolower(end($array));
-	
-	$uploaddir = 'uploads/';
-	$uploadfile = $uploaddir . 'i' . rand(1,99999).rand(1,99999) . '.' .$extension;
+
+    $validExtension = array('image/jpeg', 'image/jpg', 'image/png', 'image/gif');
+
+    if(in_array($type, $validExtension)) {
+
+        $uploaddir = 'uploads/';
+        $uploadfile = $uploaddir . 'i' . rand(1, 99999) . rand(1, 99999) . '.' . $extension;
 
 
-		if(move_uploaded_file($temp_name, $uploadfile)){
+        if (move_uploaded_file($temp_name, $uploadfile)) {
 
-			$path = $uploadfile;
+            $path = $uploadfile;
 
-			$data = ['message' => 'Je participe au concours pardon maman ! Voici ma photo :) Rejoignez le concours : https://www.facebook.com/Concours-photo-pardon-maman-Community-1926870424202660/',
-	 		 'source' => MyController::$fb->fileToUpload($path),
-			];
+            $data = ['message' => 'Voici ma photo pour le concours Pardon Maman, Rejoignez le concours : https://www.facebook.com/Concours-photo-pardon-maman-Community-1926870424202660/',
+                'source' => MyController::$fb->fileToUpload($path),
+            ];
 
-			$api = ApiController::getInstance();
+            $api = ApiController::getInstance();
 
-				try {
-					$photos = $api->postRequest('/me/photos', $data);
-				} 	catch(Facebook\Exceptions\FacebookSDKException $e) {
+            try {
+                $photos = $api->postRequest('/me/photos', $data);
+            } catch (Facebook\Exceptions\FacebookSDKException $e) {
 
-		  				echo 'Error: ' . $e->getMessage();
-		  				exit;
-					}
+                echo 'Error: ' . $e->getMessage();
+                exit;
+            }
 
             // On supprime le fichier
             unlink($uploadfile);
-			header("Location:index.php?action=participate&upload=Ok");
-}
-	 	else {
-			header("Location:index.php?action=participate&upload=Nok");
-		}
+            header("Location:index.php?action=participate&upload=Ok");
+        } else {
+            header("Location:index.php?action=participate&upload=Nok");
+        }
+
+    }else{
+        header("Location: index.php?action=participate&upload=type");
+    }
 }
